% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match.R
\name{get_match}
\alias{get_match}
\title{Scrape JSON of Match court vision data}
\usage{
get_match(year, matchid, event = c("ao", "rg"))
}
\arguments{
\item{year}{Year of match (yyyy)}

\item{matchid}{Character matchid (SM001)}

\item{event}{One of "ao" (Aussie Open) or "rg" (Roland Garros)}
}
\description{
Scrape JSON of Match court vision data
}
\examples{
get_match(2021, "SM001", "rg")
get_match(2021, "MS701", "ao")
}
