% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_points.R
\name{get_points}
\alias{get_points}
\title{Create point-level data.table for court vision match}
\usage{
get_points(x, matchid = NULL, year = NULL, event = NULL)
}
\arguments{
\item{x}{List from JSON input, return from \code{get_match}}

\item{matchid}{character matchid as on grand slam results page (e.g. 'MS001')}

\item{year}{Year of match (yyyy)}

\item{event}{Character name of event}
}
\description{
Create point-level data.table for court vision match
}
\examples{
get_points(get_match(2021, "MS701", "ao"), "MS701", 2021)
}
